refactor(PokemonList): extract Pokemon list to separate component
