+++ src/App.css
@@ -5,19 +5,19 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-  border: red solid 2px;
+  /* border: red solid 2px; */
 }
 
 .pokemonListContainer{
   height: 97vh;
-  border: solid 2px red;
+  /* border: solid 2px red; */
   overflow:hidden;
   overflow-y:scroll;
   text-align: center;
   background: var(--dark-blue);
   border-radius: 5px;
   width: 25%;
-  /* border: 1px solid var(--light-blue); */
+  border: 1px solid var(--dark-blue);
 }
 
 .pokemonListTitle{
